refactor: change folder name to kebab-case
