+++ src/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import CruiseContextProvider from './store/CruiseContextProvider';
-import CruiseHeader from './components/common/CruiseHeader';
-import CruiseMain from './components/common/CruiseMain';
+import CruiseHeader from './components/common/cruise-header';
+import CruiseMain from './components/common/cruise-main';
 
 const App: React.FC = () => (
   <Router>
