+++ src/pages/agent/agent-type-box/index.tsx
